Add logic to detect empty slots in starting lineups
In addition to players with injury statuses, another red flag
when looking for team inactivity is having empty player slots
in the starting lineup. This adds a dummy player with the ID
that Sleeper returns when a roster spot is empty to the list of
inactives, allowing the empty spots to be reported by the script.