Update status check to handle new empty status
It appears the API changed in some small way for players without a
status, potentially adding a blank status instead of returning None?
Regardless, the script now handles both a blank and an empty (but not
false according to Python) status for players and excludes those from
the standard player set.