Remove special handling of inactive offensive players
The original version of this only checked for inactive players on
offense, but after looking at things more I realized there's no reason
to exclude defensive players from the list of inactives. This removes
the special-casing that filtered out non-offensive players, making
this usable for IDP as well as standard leagues.