Improve WeeklyScore formatting past Week 9
As the season dragged on, the results for earlier weeks (with a single
digits) clash with the results from the later weeks (with double
digits). To resolve that, this changes the week formatting to right-pad
the number on the Week so that it always takes up 2 spaces, providing
more consistent formatting when looking at season-long lists.