Update starting lineup parsing to be more efficient
Previously this has been pulling each starting lineup and
then checking if each inactive player is in that lineup. This
was due to being easier (and not being great with list comprehension)
but is less efficient when the number of inactive players exceeds
the number of starters on a team.

With this refactor, we instead iterate over each starter and use
list comprehension to find the first inactive that matches a starter.