Add logic to handle matchup starters being None
While running the script this evening I found a roster
that's being reported as having a null starters list which
caused the code to crash. The team has starters specified when
looking through Sleeper's UI, so it's not clear why this is
happening.

For now, this just logs the user's league and name and then skips
their lineup, allowing the script to run to completion.