Add ability to output round.pick instead of raw adp
When outputting the ADP data the script by default shows
the average position for a player (for example a player
drafted at the 4th and 6th pick has an adp of 5). However
that's not generally how people look at draft picks, they
instead see them as what round and what pick within the round

With this change you can pass league size to the script, and
when printing human-readable outputs it will instead format the
ADPs using round and pick instead.