Add script to find latest transaction per user
Often when looking at a league we want to find out when
a user has last made a transaction. This script handles
all of the logic to find that and prints out the result
in a human-readable format.

The functions written here are also meant to be consumable
by another class if they wanted to pull in transaction data