Add script to search and display all trades
This script is similar to the league transactions script
that was recently added, but instead of looking for summary
information to find a user's last transaction this is focused
on printing all of the trades that have occurred in a league.

The output is pre-formatted for Discord with markdown but that's
easily changed in the future if desired.