Initial commit for the script refactor
This is the beginning of, for the lack of a better phrase, the
second version of these scripts. The overall goal is to build a
library that allows for a platform-agnostic script on the frontend
to decide what FF service to query. To that end there's a lot of
organizational reworking being done here.

1. Move all of the existing scripts into old/. They're needed for
   reference but will eventually all be deleted.
2. Create a libray directory under which the entire "backend" code will
   live
3. Create a model directory in the library containing the data container
   classes. These are at least a best-effort at beginning the
   implementation of the model classes, and I'm sure this will change over
   time but for now it's a start.
4. Create a base Platform interface to provide the overall definition
   for what each FF service will need to implement
5. Create the beginnings of the Sleeper API

The intent is for this to be as agnostic of OS/directory/anything as
possible so we're using relative imports to the classes within each
library. That may bite me eventually but at least as a prototype it
works.