Add new version of ADP script with Sleeper support
This is a port of the previous adp.py script using the new library
architecture. Everything in adp.py is fully platform-agnostic,
interacting only with the methods exposed by platform.py while the
Sleeper-specific implementation is abstracted away into the sleeper
platform class.

Minor tweaks were required to the model as part of this, and with this
complete I removed the test.py file used for early validation.