Add platform support for pulling Fleaflicker ADP data
With Sleeper support added in the previous commit, this was the first
main test of the architecture and whether or not it was possible to add
a second platform with the same basic interface for pulling information.
And thankfully it was, with a few minor tweaks.

Added here is the Fleaflicker platform files, both the base interface as
well as the API file. The methods for Fleaflicker matches that of the
overall Platform interface, with stark differences appearing in the API
interactions and the processing to produce the data used in the
frontend.

With that added, adp.py was modified to include a switch between the two
platforms through the CLI