Add new version of trades.py for Sleeper leagues
This is a rewrite of old/trades.py using the new architecture.

As part of this, the Sleeper backend now loads data on roster_id and
user_id/user_name into memory when the league is initially loaded. This
slow down operations as there is a lot of frontend overhead, but it
reduces the amount of programming that needs to be done in each other
method down the line for Sleeper.