Add fleaflicker support for trade script
This extends the trades.py support to Fleaflicker. By and large
the backend is the same other than one quirk of the platform that
required some accomodation in the output script.

Fleaflicker's API doesn't report what you lose in a trade, just
what you gain. So the code for outputting results needed to be
updated to only look at assets gained (in the two-person world)
and also look at if there was anything to print as far as loss/gain
in the multi-team trades.

Other than that 0 changes were made to the driving script. One minor
change was made to the TradeDetail to allow for sharing pick slot
on traded picks as that's readily available for this-year trades.