Move league filtering to the platforms
Because we're doing pre-processing on leagues when we initially
retrieve them, it behooves us to only do that on the leagues we
care about instead of pre-processing many leagues to filter it down
to a few. So this moves the filtering into the platform-level functions
to retrieve the leagues with a default "match all" argument for callers
that don't want to pass that in.