Add leaguescoring script with Fleaflicker support
This reports out weekly and season long score results, sharing the max
(and the min) weekly scores each week in a range as well as the max (and
min) season-long scores.

This was fairly straightforward other than a refactor away from the
original model proposal to have separate season and weekly score
objects. Other than that the only real change here is a slight fix to
how user data is retrieved for a league (using year-specific fetches)
and always retrieving the Fleaflicker league list from the current year,
as previous years are empty.