Add script to detect inactive rosters in Sleeper
This goes through each starting lineup to find any players who have an
inactive status (not Questionable) or are on BYE and then reports back
the results so that commissioners can follow up with league members
individually.

This is only implemented for Sleeper because I can't yet figure out how
to pull the list of starters from Fleaflicker.