Add more information to Sleeper pick trades
Sleeper's API doesn't list specific pick slots in their transaction
or trade API, so previously I've attributed picks as a "2023 1st"
instead of "2023 1.07". However, the original owner of the pick is
available on the trade, and when combined with the data in the draft API
you can attribute the pick.

So this adds a GET call to the draft board when processing trades, and
then uses that information to attribute picks in the current year.