Refactor ADP script logic into a local function
In preparation for these scripts to potentially be used
in a bot I'm writing, this refactors the logic that previously
existed in main to instead by inside of a function, that
main then calls.

This retains all of the CLI tools and arguments, ensuring that the
defaults on the CLI match the defaults in the function that others can
call into.

This also renames several of the local helper methods, to indicate that
they're not meant for external usage.