Initial commit for setting up a bot for testing
This puts together bot_test.py. It's a simple bot, built on the
discord.py library, that currently sends the results of the ADP script
to a forum channel that the user specifies. It needs some more refinement
before I can release it, but wanted to get this committed to move
it over to the desktop computer.