Add FORMATTED_CSV format for output
The existing formats are sufficient for either anaylsis usage (in the
case of CSV) or terminal usage on desktop (in the case of
HUMAN_READABLE), but lacked a middle-ground. Somewhere where the ADP
information was formatted by round/pick, but was still further parsable
for different display methods.

This is resolved by adding a third OutputFormat, FORMATTED_CSV. This
retains the CSV format, but formats the raw pick data into round.pick
format like is shown in the HUMAN_READABLE output. This can then be
consumed by downstream users (such as a discord bot wanting to reformat
the data).