Update ADP message content to use embeds instead of code block
After using this for a few days, several people have called out that the
code block formatting doesn't look good on mobile (and that it's
non-obvious why it looks bad). Ultimately want I want is a table format,
but Discord markdown doesn't support that.

So instead, I'm moving to using Discord embeds that put the player
name on one line (as the "title") and then all of the data in a
single-line code segment (for spacing) below it (as the "value"). This
creates a more pleasant viewing experience on both mobile and desktop,
without detracting from the raw data and the ability to parse it.

Since we're using embeds, I'm leveraging the embed color and matching it
to Sleeper's colors for each position, at least for the FTA posts. I
imagine more refactoring will come around once I add this for NarFFL as
well, but that's to come later.