Encapsulate trades script logic into a callable function
Similar to the recent change we made to the ADP script, this updates the
trade script to encapsulate all of the business logic in the main method
into a callable method, that is then called from main. From the script's
point of view the functionality is no different, but it has the added
benefit of allowing outside classes to call the business logic.

As a part of that, this also renames a bunch of the "private" methods to
be prefixed by an underscore.