Add bot commands to periodically post FTA trades
This leverages the recent changes to the trade script, creating a
command that polls for trades every 30 minutes. When a trade is posted,
the most recent timestamp is stored locally so that we only post new
trades (hopefully) and don't re-post existing ones.

All of this is running with a looping task, that is started and stopped
using specific commands.

Because this requires the bot to maintain state, there are two local
files under a new `bot_data` directory that store the channel to write
to as well as the last trade timestamp.