Move trade queries to a background thread
When running the trade queries I've been running into issues where they
block the bot's thread for >10 seconds, triggering warnings. To resolve
this, this commit leverages asyncio to do the queries on a background
thread instead.

Doing this allows us to loop much more frequently, so that's enabled
here as well.

Additionally this includes code to automatically start the trade posting
when the bot comes online. We'll see if I like that in practice, but for
now it's useful for testing.