Add error handling for Sleeper API errors around the main user
We've intermittently seen issues where retrieving the user fails. In a
world where this is a one-off script that's been fine, just rerun the
script. However in a world where we're running these in a polling
fashion in the background.... that's less fine.

With this change we now more gracefully handle empty responses for
retrieving the user and retrieving leagues for an empty user. I expect
there will be more hardening down the line, but this is at least a good
first step.