Add custom hashing function to Trade class
This is in preparation for better dup detection on trades, so that I can
know if we've posted a specific trade before or not.

I originally implemented this with the built in Python hash functions
and classes. But apparently an object's hash isn't consistent between
runtimes, so I can't write a hash and then have it be the same hash in a
later run. Which defeats the purpose of persisting the hash. So instead
I'm using hashlib to do an md5 hash of the trade details.

If a league has more than one trade in a given second they'll come up as
duplicates, which I'm fine with for now.