Update FTA trades commands to use trade hashes instead of timestamps
The previous trade implementation relied on trades coming out in order.
Which is fine... until Sleeper finalizes a trade without updating the
last updated timestamp on the object. Which I can't for sure know is
happening, but I'm pretty sure I've seen it happen before.

This now sidesteps that entirely be using the trade hashes. Whenever a
trade is posted, the hash (along with other descriptive information for
debugging) is written out to a file. At the top of the trade command we
read that file to know what trades have been posted, and ignore any
where the hash is already present.