Add automatically retry to API request helper
As I continue trying to chase down the consequences of intermittent 522s
from Sleeper's API, I realized that instead of trying to patch 1000
holes I could just try to fix the root cause.

With this change, every failed API request will automatically retry
once. Having already deployed this locally, the bot would previously
crash the background task within an hour without the change, but with
the change it's been running for nearly 4 hours with 2 retries that
prevented background crashes. So I'll take that as a win, and continuing
monitoring to see if there are still some holes to patch after this.