Add id field to Trade model
With us now needing an immutable identifier for a trade, it was
suggested to me that I should look at see if they have an ID on the API.
Which they do. Which seems much simpler than creating my own custom
hash.

So this adds the plumbing for an ID field and updates both platforms to
grab that. Next step is to replace the existing hashing with the trade
id.