Add file to track if we should be posting FTA trades
While doing testing for the bot, I'm often stopping and restarting the
bot which causes a lot of churn on queries for trades, because on ready
the bot always starts the periodic task for trades. There are times when
I'm doing development and I'd rather just take the trade loop down for a
while.

What this does is create a file to track a simple boolean value for
whether or not we should be posting FTA trades. On ready, trade posting
will only start if the file exists and the content is set to "True".
This flag is also set whenever you use the start/stop posting commands
so the value passed to the bot persists across restarts (and doesn't
require manual file editing of the data).