Add NarFFL ADP commands
Up to this point, everything has been for the FTA leagues. However I do
admin work for both FTA and NarFFL, so wanted to see if I can re-use
some of this as common code between both platforms.

So what this does is
 - Breaks the methods for posting ADP data into a forum into more
   generic functions
 - Stores some more common fields (like position color) into constants
 - Uses the now-generic functionality to add commands for posting NarFFL
   data

The transition didn't go quite as smooth as I'd like, because NarFFL has
100+ leagues and that presented a couple of challenges

 - The request to pull ADP data needed to be moved to it's own thread,
   as querying 132 leagues takes too long. Thankfully we were already
   doing this in the trade functions so I did the same thing here with
   asyncio
 - N being over 100 for players messed up the single-line formatting on
   mobile, so I had to rework the ADP data line slightly. Now, ADP is
   instead Av, and the number of times is wrapped in parens. At least
   for N <1000, this works fine and consistently.