Add year argument to trades in order to retrieve past years
This is more specific to Fleaflicker, where the league ID doesn't change
year to year. So when you ask for a league's trades, it gives you all of
them throughout time. Previously there was a check for "is current
year", but that doesn't play well with trying to query past trades.

So now the platform-level command to retrieve trades accepts a year
parameter, that Sleeper doesn't use but Fleaflicker does to know what
year's trade to return.