Add diagnostic command to check the status of running tasks
I've been trying to figure out how to check the status of the tasks
while I'm AFK. I can re-start them manually with other commands, but if
there's nothing happening (as there often won't be) then it's hard to
know if they need to be restarted.

This adds a small command I can run in the dev server to check the
status of the existing commands.