Refactor inactives script to leverage the team include/exclude logic
With the logic built into the backend to include or exclude specific
teams from the inactive players results, this extends the inactives
script to pass the appropriate fields to the backend. This isn't plumbed
into the CLI interface, but only into the new method that encapsulates
the functionality of the inactives script.

That method mostly exists to be called by the discord bot, which is
coming next :)