Add commands for posting FTA inactivity
This adds a few different commands, all focused on allowing the league
admins to make posts to the FTA server for league inactivity.

 - `post_inactive_fta_rosters_for_select_teams` which allows you to
   filter the inactive reports down to a subset of NFL teams (like for
   Thursday Night Football)
 - `post_inactive_fta_rosters_excluding_select_teams` which similarly
   allows you to filter out specific NFL teams (like after Thursday
   Night Football)
 - `post_fta_inactives_to_forum` whihc posts all of the reports for the
   week to a forum for commissioners, with a comment on any players that
   were ignored
 - `create_fta_league_to_channel_mapping`, which is a utility that makes
   it much easier to create the mapping of league name to Discord
   channel.

The final command is important because the first two post the reports to
the individual league channels, increasing the visibility of the reports
overall.

All of this leverages utility functions that could be used in a similar
set of NarFFL commands (or at least some of them) should we get there,
although they're not used outside of FTA commands at this point.