Update logging on all longer-running commands
I had previously added a log-line at the start of every command, but
have since realized that with some of the longer-running commands it
would be nice to know if it's done (and nothing happened) or if it's
still running.

So this just adds a log line at the end of the longer running commands
(that are doing a lot of queries across the leagues) so that glancing at
the logs provides a better picture for what's going on.