Add sleep before failed requests are retried
When Sleeper is returning 5xx responses, often sending the request back
to them just yields the same result. In order to attempt to avoid that,
before retrying a failed request we now sleep for 15 seconds.

This also removes some of the response logging, as it was unncessarily
verbose and unhelpful. Instead of trying to solve specific situations,
I'm now just adding general hardening around everything