Increase reliability of the trade tasks with try/except
Periodically the trade tasks are failing because something goes sideways
on the API side, and the resulting exception bubbles up and is
unhandled.

This change "handles" the exceptions by just accepting that the
periodic task has failed and cancelling it. The data being collected
doesn't need to be handled immediately, and can instead wait for the
next run of the periodic task.

Doing this has increased the reliability of the tasks and reduced the
need for manual intervention to 0 (so far).