Add support for inactive starter detection in Fleaflicker
Up until now, the code to retrieve and parse the starters in Fleaflicker
wasn't available in the backend library. Without that, we weren't able
to report any inactive starters.

This change adds the logic in the Fleaflicker backend to retrieve
starters and then check for inactive players in a given week. The
frontend interface for this is the same for both Sleeper and
Fleaflicker, allowing inactives.py to continue working as written.