Add scheduled task whose job it is to check other tasks
We're intermittently seeing issues where the trade tasks never finish
and just get stuck. In that case, cancelling it and restarting it is
sufficient to resolve the issue. This automates that by checking the
next scheduled start time of the task against now, and restarting the
task if the next instance should have begun.

This also throttles the FTA trade task to once in 10 minutes, to give
headroom as the leagues being checked grow.