Convert year to an int in all cases, except for API requests
Looking through the various scripts I've been inconsistent on using a
string vs an int everywhere for year. About the only consistency is that
the API files use strings, because you have to to make a string.

This converts everything to a year, and then forces the API files to do
the conversion to a string there. Years are numbers, they should be
treated as such.