Cast trade id as a string when checking for dupes
With the first Fleaflicker trades going through, I discovered that
apparently Fleaflicker trade ids are numbers and not strings. The end
result of that is that when the bot does a comparison of the trade id
(in numeric format) with the list of posted trade ids (in string
format), the trade always identifies as not yet posted.

This casts the trade id as a string in order to properly do the
comparison, and not rely on the type on the id coming from the
backend platform.