Refactor LeagueScore script to use a helper method
In preparation for using the logic here programatically, the bulk of the
script logic has been moved into a helper method. As a part of that, and
in order to reduce the number of API calls made, that helper method
returns the results in bulk in a single custom object.

An alternative approach here that was considerted but discarded was
having dedicated methods for a single set of scoring results (in lieu of
the bulk object) but that was deemed to expensive and unnecessary when a
custom object could be used.