Add flag to inactives to filter by primary user
Separate from it's use at a broad level, the inactivity script can also
be used at a micro-level to check for inactivity on a single user's
rosters. To enable this, there's a new flag that, when passed, limits
the output of the script to only reporting inactives on the specific
user being provided.

Now I can use that to check my own rosters for ianctives across all
leagues :)