Add feedback message while long-running commands run
If you're not staring at the bot logs, it's hard/impossible to know
if your long-running command is still going (and therefore everything is
fine) or if it's completed but not done anything (at which point
everything may or may not be fine).

To combat this uncertainty, the bot will now post a reply to the command
indicating that it's running, and then delete that reply once it's
complete. I'd love to user ephemeral message for this, but then I'd have
to convert everything to slash commands. Which I'm not against... but
not right now.