Add ability for bot to ping people with inactive rosters
One piece that's lacking from the inactive starter posts is the ability
for the bot to actually notify the user. We currently rely on them
seeing the message themselves or having someone else manually tag them.

With this change, users can now register their fleaflicker/sleeper
username and tie that to their Discord user on the backend. Once that
mapping exists, if that user has an inactive roster they'll be pinged
here in Discord.