Add link to the roster in inactive league posts
Unfortunately this can't be put into the name of an embedded field (so
we can't just hyperlink the team name), but this adds a link to the
roster just below the team name. This is useful in a handful of
scenarios, such as when someone is missing a starter and you want to see
if they're on the bench or just nowhere to be found, or if you just want
to double check what's going on with a specific roster.