Add commands for removing username registrations
Previously username additions were add-only, and the only way to do a
removal was to have me manually edit the data file on the server. This
is obviously less than ideal...

So now there's a command! Users can run the removal commands to delete
all registrations associated with their Discord account.