Initial POC for a bot using slash commands and cogs
When I wrote the bot last year, I did so with little/no experience with
what I was doing. I found a tutorial on how to make basic commands, and
then used that hammer to hit everything I could to achieve my goals.

A year removed, and a hopefully more refined toolkit, I want to do this
the right way. The initial file is 1800 lines and is unwieldy to use,
and the commands are verbose (intentionally), but that makes them hard
to type. Not to mention that everything is currently "secured" using
server roles which just kinda sucks.

My goal is to port the whole bot functionality over to using slash
commands, and breaking logical groups into cogs for easier editing. To
enable that, thos has a basic Bot class that invokes a Test cog with a
Hello World slash command!