Add global error handler for slash commands
Obviously errors never happen. But in cases of bit flips or other
equally anamolous behavior, it's good to check just in case...

This adds a global error handler for all app commands in the command
tree. We can also add per-command error handlers if needed, but this
serves as a nice catch-all. In the future we may want to expand to more
error types, but this serves as a good starting point for now.