Add new cog for leaderboard commands
With the basic outline of how things are structured set up for the ADP
commands previously, this just expands on the same infra by creating a
leaderboard cog containing all of the leaderboard commands across the 3
servers.

One thing to call out is I ran into a 15-minute limit on interaction
deferral when running all of the NarFFL leaderboards in series. So I
used asyncio to run them in parallel instead, which finishes inside of
15 minutes on a full NarFFL year worth of data. We'll likely port that
back to other areas in the future!