Add Cog for inactives scripts
This represents the last major lift-and-shift of the original
functionality, moving over all of the code for storing Discord:Platform
username mappings as well as the logic for retrieving inactive starters
and posting them.

The logic between the two is intertwined because ianctives are the only
place that use the Discord:Platform mapping, although that could be
moved out separately if needed.