Update Sleeper backend to cache player information
A long-standing inefficiency has existed in the Sleeper code, where we
end up retrieving new Player information every single time a Sleeper
Platform Object is instantiated. This gives us the most up-to-date
information, at the cost of an expensive API call (5 MB of data) every
single time. And in the vast majority of cases we don't need information
_that_ fresh. Sleeper's own API docs advise calling this once a day. But
fixing something that isn't broken and isn't causing problems doesn't
rise high on a priority list, which is why it's existed this way for so
long.

But the goal is that this should fix it. With this change, we are
caching player data locally on the server/file structure that's used for
the scripts. Every time a Sleeper Platform is created we check for
existence of the data file, and see if it was written less than 24 hours
ago. If it was, we serialize the data out of that instead of calling the
API.

There are some scenarios where we need more current data, such as when
we are checking for inactive players and their status just changed. So
there is an override flag in the Sleeper constructor now to force data
refresh, and we set that flag in the inactivity script/functions.