Add general-purpose depth chart command for Sleeper
This came as the result of a conversation in one of the servers, where
people wanted an easy way to view someone's roster in Sleeper as a depth
chart. Given a username and league name, this command will look up that
user's roster, pull the list of players (sorted by starters/taxi/bench)
and then print things out per position, highlighting which of the three
above categories each player fits into.

To facilitate this a couple of additional changes were necessary in the
library.

- New Roster class to pull a team's list of players and store it
- New method in Sleeper platform to retrieve a Roster
- Makes players comparable, since we need to check which list a Player
  object is in
- Adds a flag to both platforms that can be used to retrieve pre-draft
  leagues, since you might want a depth chart even if the league hasn't
  drafted yet