Add traceback printout to bot error handler
Ever since adding the error handler I've been a little frustrated,
because it seemed like a better in-Discord experience came at the
expense of losing debug logging with the stack trace.

After a little googling, it turns out I can still print the stack trace
using the traceback library. So this adds that at the end of the error
handler.