Fix error that occurs when a roster doesn't have a taxi
My code was assuming that all rosters had a taxi, when in reality if the
league doesn't have that configured then it returns a NoneType in the
`taxi` field which was causing the roster pull crash. So now we check
for the existence of a non-empty `taxi` field first.

I don't have that for starters or bench, because as far as I'm aware
every roster will have that.