Add support for retrieving a subset of matching leagues by name
Everything else up to this point has assumed that we can and should use
a regex to reduce a user's league set. This is fine when it's being done
in the backend, but I don't want to parse a user message for a regex.

This adds a separate named string parameter to allow for passing just a
string to use as the subset of a name.

The main utility of this is it short-circuits retrieving user
information for a ton of leagues when it's not needed, which saves a lot
of time in cases where a user has a lot of leagues.