Add special-case handling for dual-position DL
Sleeper lists DL who play in a 3-4 defense as LB/DL. Which is fine for
roster flexibility but less fine when they report through the scripts as
LB when they're really not.

So now, if a player has multiple positions and DL is one of them, we
will cache that player's position as DL instead of LB (which is what the
bare "position" field classified them as". The result is more accurate
data coming out of the API.