Add logic to retrieve and print future picks in depth charts
The most requested feature after adding the sleeper depth chart was the
ability to print out future picks. So this adds logic to the back end
Sleeper library to retrieve future picks (by parsing through the list of
traded picks) and also adds logic to surface that in the depth chart
commands.

As a part of this, the league object now stores whether it's redraft,
keeper, or dynasty, so that we don't print out future picks in redraft.