Retrieve rookie-draft round count from the league object
Retrieving the number of rookie draft rounds from the most recent draft
falls apart when the most recent draft is a startup, at which point you
end up returning like 30-40 rounds. The League API has a draft_rounds
settings that corresponds to the supplemental draft rounds settings, so
we can (and now do) retrieve the number of draft rounds from that
instead.