Add better error handling when looking up rosters before a draft is done
Right now we just crash because the player list from the API is empty, and
the code assumed that we could iterate over it even though it was None.

This adds better handling within the Sleeper API implementation to not try
to iterate over None, and then better user-error return when it occurs.