Add draft slot to current-year picks on depth chart
Previously all draft picks in the depth chart command were just the year
and round. This is fine for future picks, but if we're pre-draft then (hopefully)
the commish has assigned people to specific draft slots. In that case we can
include the pick and round (1.02) instead of just the round (1st).

This is leveraging the existing parsing of the draft api response that's used in
trades. So while it should be fairly robust as that's built to handle most draft
variations, using old code for a new purpose can be risky so keep an eye out for
crashes related to this.