Fix logic to build Sleeper roster from draft
When I modified the depth chart command to include specific draft picks, I
inadvertently broke the logic to build a roster from an in-progress startup
draft (and didn't realize it because startups don't happen in January usually).

When originally building the roster from a draft it didn't need to know the
id of the roster and instead passed around 0, but that no longer works. So
this change adds logic to `Sleeper.get_roster_from_draft` to retrieve the
actual roster id.

I also improved a previous bugfix I had made to separate the number of picks
in the startup draft from future rookie drafts. I can pull the number of picks
in the current draft from the draft response, and future drafts from the league,
instead of choosing one or the other (which is what I was doing).

All of this results in a better experience, and no errors, when retrieving the
depth chart from an in-progress startup draft.