Update help strings to use constants to populate defaults
Last year I moved all of the scripts over to using the libCommon
default year as the default argument, but left the actual help
strings as hardcoded numbers. In hindsight this seems silly, so I've
updated the help strings to reference the same variable so they
automatically stay in sync with the actual default value.