Add initial commands for tracking draft stats
For years there have been commissioners in FTA wondering if we can see who is
taking the most time otc for their draft. But since the Sleeper draft API
doesn't include the time of the pick in their response, it can't be
constructed after the fact.

The goal here is to have a long-running task to monitor drafts, and in
real-time monitor how long people take on the clock. This can be computed
by keeping a tally of who is currently on the clock in each loop, and then
updating time elapsed every time the loop sees a new pick.

Before we can do that, we need the ability to properly configure what drafts
are being tracked, and that's the commands added here. You can now add, remove,
and list the leagues being tracked through the bot.