Add looping task for live-monitoring and storing draft data
This builds out the core business logic of the initial draft-stats
functionality. On each loop (run every 5-6 minutes for now), the loop
checks each active draft for new picks and appropriately attributes
the last-pick time if there were any new picks. All of this is stored
to local files to be analyzed later (and the retrieval/analysis command
is yet to be written).