Remove verbose logging from update_draft_stats loop
During initial development and debugging it was good to have logs
telling me how many active drafts there are as well as logs to tell me
start/end time on the loop, but now that things are more mature that's
not needed as much. Removing these to clean up the logs.