Add raw CSV output to the FTA draft stats post command
Had someone request the raw CSV draft data so they can more easily do a
visualization over time of how things move. So now you have the option of
passing a text channel to the `send_all_fta_adp_posts` command, and that's
where it will spit out the CSV data.

As part of that, I cleaned up a little bit of the CSV data format from the
script, specifically truncating ADP to 2 decimal places and adding the
players position.