Add special handling for dual-position WRs
This is mostly the Travis Hunter commit, but in general I want to treat
players who are eligible at WR as WR when looking at the data. This works
for the purposes of the scripts because they're mostly offensive focused,
and also that most players who are eligible at WR have it as their primary
role. I may reconsider this in the future and narrow it down to just Hunter.