Add new command for printing the depth chart of an unowned team
With all of the bot commands relying on a User, when a team doesn't have
an owner you can't use the existing `sleeper_depth_chart` command to print
out the roster. This use case isn't uncommon in the offseason (as you're trying
to find new owners), and is something I wanted to handle.

This creates another sleeper-specific depth chart command that calls
directly into the Sleeper library, calling new methods to retrieve
information without a User. The end result is we can now retrieve league
information and a depth chart for a team based on IDs, not on Users.